minigame {minigame}	R Documentation

guess the random number

Description

You need to guess a number according to the requirements until you get it right, and then tell you how many times to guess.

Usage

Game()

Examples

Game()

[Package minigame version 0.1.0 Index]